lint

+++ argobytes/argobytes_brownie_project/contracts/external/uniswap/IUniswapV2Callee.sol
@@ -2,5 +2,10 @@
 pragma solidity 0.8.5;
 
 interface IUniswapV2Callee {
-    function uniswapV2Call(address sender, uint amount0, uint amount1, bytes calldata data) external;
+    function uniswapV2Call(
+        address sender,
+        uint256 amount0,
+        uint256 amount1,
+        bytes calldata data
+    ) external;
 }
